requirements.txt
pandas==2.2.2
pyarrow==22.0.0
requests==2.31.0
python-dotenv==1.0.1
duckdb==1.2.1
pytest==8.3.4
responses==0.25.6
